1. Created while loop inside "Play" selection to keep play going so long as player score is <= 21
2. Moved methods: checkName, checkChoice, and checkSelection
to Player class

3. Changed Game class from an abstract class to a regular class
as nothing was inheriting from it

4. Moved methods: checkScore, and declareWinner to Game class

5. Created callRules method in game class for all rule print
statements
