+++ SYST17796_ProjectStarterCode/src/ca/sheridancollege/project/MoveControl.java
@@ -1,0 +1,11 @@
+package ca.sheridancollege.project;
+
+/**
+ * @Rich Smith at ZenOfProgramming.com
+ */
+
+public enum MoveControl
+{
+   Hit,
+   Stand;
+}
