+++ SYST17796_ProjectStarterCode/src/ca/sheridancollege/project/Player.java
@@ -55,4 +55,44 @@ public class Player
       // WAS ORIGINALLY ABRACT b4 CHANGING PLAYER CLASS
    }
 
+   //Methods that check validity above (can be moved to appropriate classes)
+   public static boolean checkName (String name)
+   {
+      boolean validName = true;
+      // Loop to check if there are any characters that are not alphabetic
+      for (int i = 0; i < name.length(); i++) {
+         if (!(Character.isLetter(name.charAt(i)))) {
+            validName = false;
+         }
+      }
+      return validName;
+   }
+
+   public static boolean checkChoice (String choice)
+   {
+      boolean isValid = true;
+      // If statement to ensure user chooses one of the three options
+      if (!choice.equalsIgnoreCase("PLAY") && !choice.equalsIgnoreCase("RULES")
+          && !choice.equalsIgnoreCase("EXIT")) {
+         System.out.println("Error! Please choose a correct option!");
+         System.out.println("Type PLAY to start a new game. \nType RULES for the "
+             + "rules of Black Jack. \nType EXIT to terminate the game.");
+         isValid = false;
+      }
+      return isValid;
+   }
+
+   // Check Hit or Stand selection
+   public static boolean checkSelection (String playerSelection)
+   {
+      boolean isValid = true;
+      if (!playerSelection.equalsIgnoreCase(MoveControl.Hit.name())
+          && !playerSelection.equalsIgnoreCase(MoveControl.Stand.name())) {
+         System.out.println("Error! Please make a valid selection!");
+         System.out.println("Please choose to either HIT or STAND: ");
+         isValid = false;
+      }
+      return isValid;
+   }
+
 }
